Add StandardCodec
