+++ tablet/src/main/scala/wvlet/airframe/tablet/msgpack/MessageCodec.scala
@@ -1,60 +1,59 @@
 package wvlet.airframe.tablet.msgpack
 
-import java.io.File
-
-import org.msgpack.core.{MessagePacker, MessageUnpacker}
+import org.msgpack.core.{MessageFormatException, MessagePacker, MessageUnpacker}
+import org.msgpack.value.Value
+import wvlet.airframe.tablet.msgpack.StandardCodec.ObjectCodec
+import wvlet.surface.Surface
 import wvlet.surface.reflect.RuntimeSurface
-import wvlet.surface.{Primitive, Surface}
-import wvlet.surface
+
+import scala.util.{Failure, Success, Try}
 
 trait MessageCodec[@specialized(Int, Long, Float, Double, Boolean, Byte) A] {
   def pack(p: MessagePacker, v: A)
   def unpack(u: MessageUnpacker, v: MessageHolder)
+
+  def u(u: MessageUnpacker, v: MessageHolder) {
+    Try(unpack(u, v)) match {
+      case Success(_) =>
+      // do nothing
+      case Failure(e) =>
+        e match {
+          case me: MessageFormatException =>
+          // On MessagePack related error, we cannot continue unpacking
+
+          case e: Throwable =>
+        }
+    }
+  }
+}
+
+trait MessageValueCodec[A] {
+  def pack(v: A): Value
+  def unpack(v: Value): A
 }
 
 object MessageCodec {
 
   import scala.reflect.runtime.{universe => ru}
 
-  val primitiveCodecs: Map[Surface, MessageCodec[_]] = Map(
-    Primitive.Int              -> IntCodec,
-    Primitive.Long             -> LongCodec,
-    Primitive.Float            -> FloatCodec,
-    Primitive.Double           -> DoubleCodec,
-    Primitive.Boolean          -> BooleanCodec,
-    Primitive.String           -> StringCodec,
-    Primitive.Byte             -> ByteCodec,
-    Primitive.Short            -> ShortCodec,
-    surface.of[Array[Int]]     -> IntArrayCodec,
-    surface.of[Array[Long]]    -> LongArrayCodec,
-    surface.of[Array[Float]]   -> FloatArrayCodec,
-    surface.of[Array[Double]]  -> DoubleArrayCodec,
-    surface.of[Array[Boolean]] -> BooleanArrayCodec,
-    surface.of[Array[String]]  -> StringArrayCodec,
-    surface.of[Array[Byte]]    -> ByteArrayCodec,
-    surface.of[File]           -> FileCodec,
-    // TODO CharCodec, ShortArrayCodec, CharArrayCodec
-  )
-
-  def default = new CodecFactory(primitiveCodecs)
+  def default = new CodecFactory(StandardCodec.primitiveCodecs)
 
   def of[A: ru.TypeTag]: MessageCodec[A] = default.of[A]
 
   class CodecFactory(knownCodecs: Map[Surface, MessageCodec[_]]) {
-    private var seen = Map.empty[Surface, MessageCodec[_]]
+    private var seen = Set.empty[Surface]
 
     def ofSurface(surface: Surface): MessageCodec[_] = {
       if (knownCodecs.contains(surface)) {
         knownCodecs(surface)
       } else if (seen.contains(surface)) {
-        seen(surface)
+        throw new IllegalArgumentException(s"Codec for recursive types is not supported: ${surface}")
       } else {
+        seen += surface
         val codecs = for (p <- surface.params) yield {
           ofSurface(p.surface)
         }
-        val newCodec = ObjectCodec(surface, codecs.toIndexedSeq)
-        seen += surface -> newCodec
-        newCodec
+        ObjectCodec(surface, codecs.toIndexedSeq)
       }
     }
 
@@ -63,5 +62,4 @@ object MessageCodec {
       ofSurface(surface).asInstanceOf[MessageCodec[A]]
     }
   }
-
 }
